requirements.txt
Flask
Flask-Cors
psycopg2-binary 
pandas
python-dotenv
gunicorn
pgvector
bcrypt
